Updating where the proxy happens
