+++ clickhouse-jdbc/src/test/java/com/clickhouse/jdbc/ClickHouseDataSourceTest.java
@@ -21,7 +21,7 @@ import com.clickhouse.client.config.ClickHouseDefaults;
 public class ClickHouseDataSourceTest extends JdbcIntegrationTest {
     @Test(groups = "integration")
     public void testHighAvailabilityConfig() throws SQLException {
-        if (isCloud() || VersionSelectingDriver.isV2()) return; //TODO: testHighAvailabilityConfig - Revisit, see: https://github.com/ClickHouse/clickhouse-java/issues/1747
+        if (isCloud() || ClickHouseDriver.isV2()) return; //TODO: testHighAvailabilityConfig - Revisit, see: https://github.com/ClickHouse/clickhouse-java/issues/1747
         String httpEndpoint = getEndpointString();
         String grpcEndpoint = "grpc://" + getServerAddress(ClickHouseProtocol.GRPC) + "/";
         String tcpEndpoint = "tcp://" + getServerAddress(ClickHouseProtocol.TCP) + "/";
@@ -40,7 +40,7 @@ public class ClickHouseDataSourceTest extends JdbcIntegrationTest {
     @Test // (groups = "integration")
     @Ignore
     public void testMultiEndpoints() throws SQLException {
-        if (isCloud() || VersionSelectingDriver.isV2()) return; //TODO: testMultiEndpoints - Revisit, see: https://github.com/ClickHouse/clickhouse-java/issues/1747
+        if (isCloud() || ClickHouseDriver.isV2()) return; //TODO: testMultiEndpoints - Revisit, see: https://github.com/ClickHouse/clickhouse-java/issues/1747
         String httpEndpoint = getEndpointString();
         String grpcEndpoint = "grpc://" + getServerAddress(ClickHouseProtocol.GRPC) + "/";
         String tcpEndpoint = "tcp://" + getServerAddress(ClickHouseProtocol.TCP) + "/";
@@ -84,17 +84,17 @@ public class ClickHouseDataSourceTest extends JdbcIntegrationTest {
         String params = String.format("?%s=%s&%s=%d&%s", ClickHouseClientOption.CLIENT_NAME.getKey(), clientName,
                 ClickHouseClientOption.MAX_EXECUTION_TIME.getKey(), maxExecuteTime, JdbcConfig.PROP_CONTINUE_BATCH);
 
-        for (ClickHouseDataSource ds : new ClickHouseDataSource[] {
-                new ClickHouseDataSource(url, properties),
-                new ClickHouseDataSource(urlWithCredentials, properties),
-                new ClickHouseDataSource(url + params),
-                new ClickHouseDataSource(urlWithCredentials + params),
+        for (DataSourceV1 ds : new DataSourceV1[] {
+                new DataSourceV1(url, properties),
+                new DataSourceV1(urlWithCredentials, properties),
+                new DataSourceV1(url + params),
+                new DataSourceV1(urlWithCredentials + params),
         }) {
             for (ClickHouseConnection connection : new ClickHouseConnection[] {
                     ds.getConnection("default", getPassword()),
-                    new ClickHouseDriver().connect(url, properties),
-                    new ClickHouseDriver().connect(urlWithCredentials, properties),
-                    new ClickHouseDriver().connect(urlWithCredentials + params, new Properties()),
+                    new DriverV1().connect(url, properties),
+                    new DriverV1().connect(urlWithCredentials, properties),
+                    new DriverV1().connect(urlWithCredentials + params, new Properties()),
                     (ClickHouseConnection) DriverManager.getConnection(url, properties),
                     (ClickHouseConnection) DriverManager.getConnection(urlWithCredentials, properties),
                     (ClickHouseConnection) DriverManager.getConnection(urlWithCredentials + params),
